validator_backend/requirements.txt
google-generativeai
Flask
PyPDF2
flask-cors